requirements.txt
argon2-cffi==16.3.0
bcrypt==3.1.4
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.3
django-bootstrap3==9.0.0
django-braces==1.11.0
django-debug-toolbar==1.8
django-filter==1.1.0
djangorestframework==3.7.3
Faker==0.8.5
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.9
MarkupSafe==1.0
misaka==2.1.0
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
selenium==3.7.0
six==1.11.0
sqlparse==0.2.4
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
